apply asyncfeature on yolov8 ios
