+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/Assets.xcassets/AppIcon.appiconset/Contents.json
