+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/Assets.xcassets/AppIcon.appiconset/ZETIC_gray_square 1024.png
