+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/CocoConfig.swift
