+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/Vision/CameraMode.swift
@@ -1,0 +1,4 @@
+public enum CameraMode : Int {
+    case frame
+    case image
+}
