+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/Vision/CameraPreview.swift
@@ -1,0 +1,14 @@
+import SwiftUI
+import AVFoundation
+
+struct CameraPreview: UIViewRepresentable {
+    let session: AVCaptureSession
+    
+    func makeUIView(context: Context) -> PreviewView {
+        return PreviewView(session: session)
+    }
+    
+    func updateUIView(_ uiView: PreviewView, context: Context) {
+        // Update view if needed
+    }
+}
