+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/ZeticMLange.framework/_CodeSignature/CodeResources
@@ -1,0 +1,192 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Info.plist</key>
+		<data>
+		l9sxFkguH+6N2IhPSbwXrgeuFro=
+		</data>
+		<key>Modules/ZeticMLange.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo</key>
+		<data>
+		2BBPsdyARoMIIRjZAQianN7spHI=
+		</data>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
+		<data>
+		42MF3M9bhzKPN/3xJkX0EMnkCAE=
+		</data>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
+		<data>
+		PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
+		</data>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
+		<data>
+		48LzpYNIYTiZbgdXeQaHNYK0aXs=
+		</data>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
+		<data>
+		PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
+		</data>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
+		<data>
+		V9w0JZdwB0/aQNlGQH8wC9NCzXA=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Modules/ZeticMLange.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			2BBPsdyARoMIIRjZAQianN7spHI=
+			</data>
+			<key>hash2</key>
+			<data>
+			3CY6Za8hZQyCiNMdibJwFWiNl0nI6uDlLEadFqphnFQ=
+			</data>
+		</dict>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			42MF3M9bhzKPN/3xJkX0EMnkCAE=
+			</data>
+			<key>hash2</key>
+			<data>
+			NhfyCiQCxMBsD1LUx+3cgDE5l5A3rGp4EHS1QQqloeI=
+			</data>
+		</dict>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
+			</data>
+			<key>hash2</key>
+			<data>
+			eUJF0F4YispGrblZ0O8QIJ+KMj5PoL1IHsvTaA7rU/w=
+			</data>
+		</dict>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			48LzpYNIYTiZbgdXeQaHNYK0aXs=
+			</data>
+			<key>hash2</key>
+			<data>
+			nnKIs83YS50RLRDOhnU3MdBSUz5tfKr7SY0ANK2Z90c=
+			</data>
+		</dict>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			PNtZ0Ch2IVHfl1wV+lvr8X8zGQM=
+			</data>
+			<key>hash2</key>
+			<data>
+			eUJF0F4YispGrblZ0O8QIJ+KMj5PoL1IHsvTaA7rU/w=
+			</data>
+		</dict>
+		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
+		<dict>
+			<key>hash</key>
+			<data>
+			V9w0JZdwB0/aQNlGQH8wC9NCzXA=
+			</data>
+			<key>hash2</key>
+			<data>
+			DGndV2QAFryCzaH7TpRk4nezLhiJyhxfvFnjFihRm8s=
+			</data>
+		</dict>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Base\.lproj/</key>
+		<dict>
+			<key>weight</key>
+			<real>1010</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
