+++ yolov8/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8-iOS/ZeticMLangeYOLOv8_iOSApp.swift
@@ -1,0 +1,11 @@
+
+import SwiftUI
+
+@main
+struct ZeticMLangeYOLOv8_iOSApp: App {
+    var body: some Scene {
+        WindowGroup {
+            YOLOv8View()
+        }
+    }
+}
