--- yolov8/ZeticMLangeYolov8-iOS/ZeticMLangeYolov8-iOS.xcodeproj/xcuserdata/gimdoyeob.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "A266AEFB-D915-4E6A-AC9D-869EE5603EC4"
-   type = "1"
-   version = "2.0">
-</Bucket>
