--- yolov8/ZeticMLangeYolov8-iOS/ZeticMLangeYolov8-iOS/View/PreviewView.swift
@@ -1,44 +1,0 @@
-import SwiftUI
-import AVFoundation
-
-class PreviewView: UIView {
-    private var captureSession: AVCaptureSession
-    
-    init(session: AVCaptureSession) {
-        self.captureSession = session
-        super.init(frame: .zero)
-        
-        setupLayer()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setupLayer() {
-        let previewLayer = AVCaptureVideoPreviewLayer(session: captureSession)
-        previewLayer.videoGravity = .resizeAspectFill
-        layer.addSublayer(previewLayer)
-        previewLayer.frame = self.layer.bounds
-    }
-    
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        // Update preview layer frame when view layout changes
-        if let previewLayer = layer.sublayers?.first as? AVCaptureVideoPreviewLayer {
-            previewLayer.frame = self.layer.bounds
-        }
-    }
-}
-
-struct CameraPreview: UIViewRepresentable {
-    let session: AVCaptureSession
-    
-    func makeUIView(context: Context) -> PreviewView {
-        return PreviewView(session: session)
-    }
-    
-    func updateUIView(_ uiView: PreviewView, context: Context) {
-        // Update view if needed
-    }
-}
