--- yolov8/ZeticMLangeYolov8-iOS/ZeticMLangeYolov8-iOS/ZeticMLange.framework/_CodeSignature/CodeResources
@@ -1,192 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>files</key>
-	<dict>
-		<key>Info.plist</key>
-		<data>
-		DmeGIuUdl25zskAqfnQU/+LqyVE=
-		</data>
-		<key>Modules/ZeticMLange.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo</key>
-		<data>
-		gJwxG0Xt44tlyaKpnbrm9Sl7v9c=
-		</data>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
-		<data>
-		4vl9tmFtEUcxpj7K7EcYWsug4rU=
-		</data>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
-		<data>
-		KWwgsXzajmOeaco9Yr14rSIkJc0=
-		</data>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
-		<data>
-		JOZrhIIJGTiDv4EMTDeDOUmnWyY=
-		</data>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
-		<data>
-		KWwgsXzajmOeaco9Yr14rSIkJc0=
-		</data>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
-		<data>
-		6/d1diA9tFxIGvkJjRX4RYIiQK4=
-		</data>
-	</dict>
-	<key>files2</key>
-	<dict>
-		<key>Modules/ZeticMLange.swiftmodule/Project/arm64-apple-ios.swiftsourceinfo</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			gJwxG0Xt44tlyaKpnbrm9Sl7v9c=
-			</data>
-			<key>hash2</key>
-			<data>
-			01sUesH90sKqZJ5HqPCsGAz4Da7hbloSxuLs2Q+LPpg=
-			</data>
-		</dict>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.abi.json</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			4vl9tmFtEUcxpj7K7EcYWsug4rU=
-			</data>
-			<key>hash2</key>
-			<data>
-			cC0HVXr7FYm8VXMHzU1bGBBsuqIgl+3GiqhOXtNmPmk=
-			</data>
-		</dict>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.private.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			KWwgsXzajmOeaco9Yr14rSIkJc0=
-			</data>
-			<key>hash2</key>
-			<data>
-			amDrrxNO+YyyBFONwqIIUWX0+L0Y+Jm0AzwyhiQjsJA=
-			</data>
-		</dict>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftdoc</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			JOZrhIIJGTiDv4EMTDeDOUmnWyY=
-			</data>
-			<key>hash2</key>
-			<data>
-			kpC4WBy/zUfWGLYYqAM4g4TZGcGT5olI39wXtYXktEI=
-			</data>
-		</dict>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftinterface</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			KWwgsXzajmOeaco9Yr14rSIkJc0=
-			</data>
-			<key>hash2</key>
-			<data>
-			amDrrxNO+YyyBFONwqIIUWX0+L0Y+Jm0AzwyhiQjsJA=
-			</data>
-		</dict>
-		<key>Modules/ZeticMLange.swiftmodule/arm64-apple-ios.swiftmodule</key>
-		<dict>
-			<key>hash</key>
-			<data>
-			6/d1diA9tFxIGvkJjRX4RYIiQK4=
-			</data>
-			<key>hash2</key>
-			<data>
-			uIB46LQRBKjjfRjb/dsTvxc6qMdblSHOso8m22XuYHE=
-			</data>
-		</dict>
-	</dict>
-	<key>rules</key>
-	<dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^version.plist$</key>
-		<true/>
-	</dict>
-	<key>rules2</key>
-	<dict>
-		<key>.*\.dSYM($|/)</key>
-		<dict>
-			<key>weight</key>
-			<real>11</real>
-		</dict>
-		<key>^(.*/)?\.DS_Store$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>2000</real>
-		</dict>
-		<key>^.*</key>
-		<true/>
-		<key>^.*\.lproj/</key>
-		<dict>
-			<key>optional</key>
-			<true/>
-			<key>weight</key>
-			<real>1000</real>
-		</dict>
-		<key>^.*\.lproj/locversion.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>1100</real>
-		</dict>
-		<key>^Base\.lproj/</key>
-		<dict>
-			<key>weight</key>
-			<real>1010</real>
-		</dict>
-		<key>^Info\.plist$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^PkgInfo$</key>
-		<dict>
-			<key>omit</key>
-			<true/>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^embedded\.provisionprofile$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-		<key>^version\.plist$</key>
-		<dict>
-			<key>weight</key>
-			<real>20</real>
-		</dict>
-	</dict>
-</dict>
-</plist>
