--- yolov8/ZeticMLangeYolov8-iOS/ZeticMLangeYolov8-iOS/ZeticMLangeYolov8_iOSApp.swift
@@ -1,11 +1,0 @@
-
-import SwiftUI
-
-@main
-struct ZeticMLangeYolov8_iOSApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
